finalized endpoints

+++ backend/Dhbw positioning System Backend/Model/MeasurementEntity.cs
@@ -9,7 +9,7 @@ namespace Dhbw_positioning_System_Backend.Model
         public long MeasurementId { get; set; }
         public string Ssid { get; set; }
         public string Mac { get; set; }
-        public double Rssi { get; set; }
+        public long Rssi { get; set; }
 
         public virtual Measurement Measurement { get; set; }
     }
